requirements.txt
streamlit
google-generativeai
ddgs
feedparser
wikipedia
python-dotenv
reportlab
